+++ librubyfmt/src/lib.rs
@@ -12,7 +12,7 @@ pub type RawStatus = i64;
 
 #[macro_use]
 mod ruby;
-mod breakable_entry;
+mod render_targets;
 mod comment_block;
 mod de;
 mod delimiters;
