+++ librubyfmt/src/line_tokens.rs
@@ -1,4 +1,4 @@
-use crate::breakable_entry::{BreakableEntry, ConvertType};
+use crate::render_targets::{BreakableEntry, ConvertType};
 
 #[derive(Debug, Clone)]
 pub enum LineToken {
